fixed thesisProposalParticipant deletion for internal advisors

+++ src/main/java/org/fenixedu/academic/thesis/domain/ThesisProposalParticipant.java
@@ -67,7 +67,9 @@ public class ThesisProposalParticipant extends ThesisProposalParticipant_Base {
         this.setUser(null);
         this.setThesisProposal(null);
 
-        this.getExternalUser().delete();
+        if (this.getExternalUser() != null) {
+            this.getExternalUser().delete();
+        }
 
         deleteDomainObject();
     }
